# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM La Bercianita API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /agent-runs:
    get:
      operationId: getAgentRuns
      tags: [agentRuns]
      summary: List agent runs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of agent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
    post:
      operationId: createAgentRun
      tags: [agentRuns]
      summary: Create agent run log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
...
  /agent/leads/upsert:
    post:
      operationId: agentUpsertLead
      tags: [agent]
      summary: Create or update a lead (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200":
          description: Upserted lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /agent/leads/bulk-upsert:
    post:
      operationId: agentBulkUpsertLeads
      tags: [agent]
      summary: Bulk create or update leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadUpsert"
      responses:
        "200":
          description: Bulk upsert result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpsertResult"
  /agent/leads:
    get:
      operationId: agentGetLeads
      tags: [agent]
      summary: Search leads (agent)
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: business_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /agent/leads/{id}:
    get:
      operationId: agentGetLead
      tags: [agent]
      summary: Get lead by ID (agent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: agentUpdateLead
      tags: [agent]
      summary: Update lead status (agent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /agent/contacts:
    post:
      operationId: agentCreateContact
      tags: [agent]
      summary: Create a contact (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /agent/email-drafts:
    post:
      operationId: agentCreateEmailDraft
      tags: [agent]
      summary: Create email draft (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDraftInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /agent/events:
    post:
      operationId: agentCreateEvent
      tags: [agent]
      summary: Create outreach event (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachEvent"
  /agent/replies:
    post:
      operationId: agentCreateReply
      tags: [agent]
      summary: Create a reply (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reply"
  /agent/tasks:
    post:
      operationId: agentCreateTask
      tags: [agent]
      summary: Create a task (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /agent/suppression:
    get:
      operationId: agentGetSuppression
      tags: [agent]
      summary: Get suppression list (agent)
      responses:
        "200":
          description: Suppression list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuppressionEntry"
    post:
      operationId: agentCreateSuppression
      tags: [agent]
      summary: Add to suppression list (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuppressionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppressionEntry"
  /agent/runs:
    post:
      operationId: agentCreateRun
      tags: [agent]
      summary: Create agent run log (agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
...
    Lead:
      type: object
      required: [id, businessName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        businessType:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        sourceNotes:
          type: ["string", "null"]
        fitScore:
          type: ["integer", "null"]
        fitReason:
          type: ["string", "null"]
        consentStatus:
          type: ["string", "null"]
        crmStatus:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        lastContactedAt:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    LeadInput:
      type: object
      required: [businessName]
      properties:
        businessName:
          type: string
        businessType:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        instagram:
          type: string
        sourceUrl:
          type: string
        sourceNotes:
          type: string
        fitScore:
          type: integer
        fitReason:
          type: string
        consentStatus:
          type: string
        crmStatus:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        lastContactedAt:
          type: string
        nextActionAt:
          type: string
    LeadUpdate:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        instagram:
          type: string
        sourceUrl:
          type: string
        sourceNotes:
          type: string
        fitScore:
          type: integer
        fitReason:
          type: string
        consentStatus:
          type: string
        crmStatus:
          type: string
        priority:
          type: string
        assignedTo:
          type: string
        lastContactedAt:
          type: string
        nextActionAt:
          type: string
    Contact:
      type: object
      required: [id, leadId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ContactInput:
      type: object
      required: [leadId, name]
      properties:
        leadId:
          type: integer
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        sourceUrl:
          type: string
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        sourceUrl:
          type: string
    EmailDraft:
      type: object
      required: [id, leadId, subject, body, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        subject:
          type: string
        body:
          type: string
        status:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EmailDraftInput:
      type: object
      required: [leadId, subject, body]
      properties:
        leadId:
          type: integer
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        createdBy:
          type: string
...
    OutreachEvent:
      type: object
      required: [id, leadId, channel, direction, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        channel:
          type: string
        direction:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        eventAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OutreachEventInput:
      type: object
      required: [leadId, channel, direction]
      properties:
        leadId:
          type: integer
        channel:
          type: string
        direction:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        eventAt:
          type: string
...
    Reply:
      type: object
      required: [id, leadId, body, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        fromEmail:
          type: ["string", "null"]
        body:
          type: string
        intent:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
        receivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ReplyInput:
      type: object
      required: [leadId, body]
      properties:
        leadId:
          type: integer
        fromEmail:
          type: string
        body:
          type: string
        intent:
          type: string
        sentiment:
          type: string
        receivedAt:
          type: string
    Task:
      type: object
      required: [id, leadId, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [leadId, title]
      properties:
        leadId:
          type: integer
        title:
          type: string
        description:
          type: string
        taskType:
          type: string
        dueDate:
          type: string
        status:
          type: string
...
    SuppressionEntry:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        domain:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
    SuppressionInput:
      type: object
      properties:
        email:
          type: string
        domain:
          type: string
        reason:
          type: string
...
    AgentRun:
      type: object
      required: [id, agentName, createdAt]
      properties:
        id:
          type: integer
        agentName:
          type: string
        runType:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        errors:
          type: ["string", "null"]
        createdAt:
          type: string
    AgentRunInput:
      type: object
      required: [agentName]
      properties:
        agentName:
          type: string
        runType:
          type: string
        status:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: string
        summary:
          type: string
        errors:
          type: string
...
    BulkLeadUpsert:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadInput"
    BulkUpsertResult:
      type: object
      required: [created, updated, skipped, errors]
      properties:
        created:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string