# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM La Bercianita API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRun {
  id: number;
  agentName: string;
  /** @nullable */
  runType?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  errors?: string | null;
  createdAt: string;
}