# lib/api-zod/src/generated/types/bulkUpsertResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM La Bercianita API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BulkUpsertResult {
  created: number;
  updated: number;
  skipped: number;
  errors: string[];
}